various bug fixes from dev branch

+++ nyx/auxiliary_buffer.c
@@ -187,7 +187,7 @@ void set_pt_overflow_auxiliary_result_buffer(auxilary_buffer_t* auxilary_buffer)
   VOLATILE_WRITE_8(auxilary_buffer->result.pt_overflow, 1);
 }
 
-void set_exec_done_auxiliary_result_buffer(auxilary_buffer_t* auxilary_buffer, uint8_t sec, uint32_t usec, uint32_t num_dirty_pages){
+void set_exec_done_auxiliary_result_buffer(auxilary_buffer_t* auxilary_buffer, uint32_t sec, uint32_t usec, uint32_t num_dirty_pages){
   VOLATILE_WRITE_8(auxilary_buffer->result.exec_done, 1);
 
   VOLATILE_WRITE_32(auxilary_buffer->result.runtime_sec, sec);
